+++ src/Drivebase.hpp
@@ -15,7 +15,7 @@ struct Motor {
 	int ID;
 	struct PID pid;
 	struct Error error;
-	double speed { 0.0 };
+	float speed { 0.0 };
 	int32_t last_ticks { 0 };
 	long int last_time_ms { 0 };
 };
@@ -23,7 +23,7 @@ struct Drivebase {
 	struct Motor left;
 	struct Motor right;
 
-	double x{ 0.75 }, y { 0.25 }; // Position in meters, from the bottom left corner of the field
+	float x{ 0.75 }, y { 0.25 }; // Position in meters, from the bottom left corner of the field
 };
 struct DrivebaseState {
     p28::mt::Vec2 pos; // Position in m
@@ -49,23 +49,23 @@ struct DrivebaseActionState {
 };
 
 // Conversion for encoders to distance (meters)
-double ticks_to_dist(int32_t ticks);
-double accel_dist(double accel, double target_speed);
+float ticks_to_dist(int32_t ticks);
+float accel_dist(float accel, float target_speed);
 
 // Functions that move the robot must return 
 // the part of the robot that has been moved
 // with a modified state
 
 
-struct Motor get_motor_speed(struct Motor motor, double delta_s);
-struct Motor update_motor_at_speed(struct Motor motor, double speed, long int time_ms);
+struct Motor get_motor_speed(struct Motor motor, float delta_s);
+struct Motor update_motor_at_speed(struct Motor motor, float speed, long int time_ms);
 
 
 struct Drivebase zero_all(struct Drivebase drvb);
 struct Drivebase set_motorTime(struct Drivebase drvb, long int time_ms);
 
 
-// double velocity_profile(double target_vel, double dist_to_travel, double current_dist, double time_since_start);
+// float velocity_profile(float target_vel, float dist_to_travel, float current_dist, float time_since_start);
 
 }
